feat(suiteheader): Separating the data fetching function from the useSuiteHeaderData hook so that it is also possible to import the plain function that calls all the required APIs without the need to use React Hooks.
